API - routes - params error fixed
