+++ app/api/posts/[id]/delete/route.ts
@@ -1,8 +1,11 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
 
-export async function DELETE(_: Request, context: { params: { id: string } }) {
-  const { id } = await context.params;
+export async function DELETE(
+  req: Request,
+  { params }: { params: Promise<{ id: string }> }
+) {
+  const { id } = await params;
 
   try {
     const deleted = await prisma.post.delete({
