+++ app/api/posts/[id]/edit/route.ts
@@ -2,8 +2,11 @@ import { prisma } from "@/lib/prisma";
 import { slugify } from "@/lib/utils";
 import { NextResponse } from "next/server";
 
-export async function PUT(req: Request, context: { params: { id: string } }) {
-  const { id } = await context.params;
+export async function PUT(
+  req: Request,
+  { params }: { params: Promise<{ id: string }> }
+) {
+  const { id } = await params;
 
   try {
     const body = await req.json();
